Add many improvements
NOTE: this is a combination of 29 commits with the goal of removing
domain-specific applications out of the repository.

------
Rearrange the order of directories

In order to more easily include integrate native dash applications, some
rearrangments were done:

- Now all apps (dash native and non native) have to be inside
  app/dashboard.
  - In the case of native dash apps, you must add a file with the same
    name as the directory (Dash_App1/Dash_App1.py, for example). This
    file must define a Add_Dash() function (see
    app/dashboard/Dash_App2/Dash_App2.py to have an idea what you need).
  - In the case of non dash apps, you need to have a generate_html.py
    file inside the directory, which generates the output of the
    application to be shown in the dashboard.

  - Due to this change, all apps in html-apps were moved to
    app/dashboard.

- Rename html-apps to jinja_templates. This directory only has template
  files for automatic adding apps to views and __init__
- Edit generate_html.py according to the directory changes.

-------
Removes unnecesary files from the app

These files may be needed to run the app locally but aren't needed for
the dashboard, so they are discarded.

-------
Modify path of input files for non dash apps

Instead of happing to prepending "../../../" to input files for non dash
apps, now you can write the normal path and assuming the data is in the
root of the project, it should work.

Before, if you didn't have a data/mapa_incidencia/ directory, you were
shown an error until you created the folder.

-------
Add guide on how to integrate non dash apps (ES)

Also updates the readme to link to the guide.

-------
Adds guide for dash application (ES)

-------
Refactor guides into smaller chunks

-------
tweaks to html generation stage

-------
Refactor application creation

Many changes in implementation, but for developers the main changes are:
- Added script to automate the generation of files to create a dash app.
- Updated documentation for adding new dash apps.
- Redesigned navbar.
- Applications use the theme defined in the project (superhero for now,
  but it can be changed).

-------
Updates README

Updates instructions to stop the app and add notes on what command to
use to run the project.